Full support for relays.
Essentially placed around the orbit of the moon, allowing for
communications with the drones all around the system. In the event that
the relays for your team are destroyed, your drones will experience a
drift through the eclipsed parts of the board, where the are highly
vulnerable to attack from enemy ships (i.e. 1 v 5 will succeed).